Firmware v4
In this update we changed the firmware design by adding delays to the displaysroutines spliting parts of the displaying (only the display with the changed bits gets an update). This also addresses the bugs caused by the byte-array-storage (we went back to Strings).

Note: Only the AP1 module has been implemented (yet), due to unfinished designchoises regarding the decimal point in the com/nav modules.

It also hasn't been tested (yet), but I'm up to it.